Move HigherOrderCollectionProxy class.
