+++ src/Illuminate/Collect/Collection.php
@@ -11,7 +11,7 @@ use ArrayIterator;
 use CachingIterator;
 use JsonSerializable;
 use IteratorAggregate;
-use Illuminate\Support\Traits\Macroable;
+use Illuminate\Support\Traits\Macroable; // Todo refactor
 use Illuminate\Contracts\Support\Jsonable;
 use Symfony\Component\VarDumper\VarDumper;
 use Illuminate\Contracts\Support\Arrayable;
